% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{APAdotplot}
\alias{APAdotplot}
\title{Dot plot of selected APA genes}
\usage{
APAdotplot(ECoeffSig_Mat, FDR_P_cutoff = 0.05, CoeffCutoff = log(2), APAgenes)
}
\arguments{
\item{ECoeffSig_Mat}{significant APA table returned from \code{APAtest}.}

\item{FDR_P_cutoff}{The cutoff for FDR-controlled P values of Wald tests. Default to 0.05.}

\item{CoeffCutoff}{The cutoff for estimated coefficients of logistic regression.}

\item{APAgenes}{Character vector, a list of gene IDs user would like to display in the dot plot. The format should be consistent with geneID part of rownames in the significant APA table, transcript IDs, chrIDs, and all other parts of APA ID should be removed. For better interpretation, gene IDs will be converted to gene symbols in the plot.}
}
\value{
\code{APAdotplot} returns a dot plot showing the status of 3' UTR legnthening or shortening for user selected genes.
}
\description{
Draw dot plot of cluster-specific APA genes selected by user. The size of dots shows the deviation of the proportion of long isoforms from grand mean of all transcripts. The color of dots shows the direction of 3' UTR processing (lengthening or shortening).
}
\examples{
APAdotplot(ECoeffSig_Mat, FDR_P_cutoff=0.05, CoeffCutoff=log(2), APAgenes=str_extract(rownames(ECoeffSig_Mat), "ENSMUSG[:digit:]*"))

}
