% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSig.R
\name{APAtest}
\alias{APAtest}
\title{Estimate significance of APA event at gene level}
\usage{
APAtest(countMatrix, coverageCutoff = 20, ORcutoff = 0.25, adPval = 0.05)
}
\arguments{
\item{countMatrix}{Isoform-specific count matrix returned by \code{readinPAsites}.}

\item{coverageCutoff}{The cutoff for coverage of each gene (sum of long and short isoforms of all clusters). Default value is 20.}

\item{ORcutoff}{The cutoff for odds ratio of long isoform from each cluster against the grand mean of all clusters. Default to 0.25.}

\item{adPval}{The cutoff for FDR-controlled P values. Default to 0.05.}
}
\value{
A data.frame containing significant APA genes with 1) their gene-level FDR controlled P values. 2) estimated coefficients (log(OR)), P values from Wald test, SE, and OR for each gene.
}
\description{
\code{APAtest} estimates significance of APA event at gene level.
}
\details{
There are two criteria to be met to call a gene to be APA. 1) FDR-controlled p value < 0.05. 2) At least one cluster whose odds of long isoform divided by odds of long isoform of all clusters > OR cutoff.
}
\examples{
sigAPA_table  <- APAtest(countMatrix=Matix_returned_from_readin, coverageCutoff = 20, ORcutoff = 0.25, adPval = 0.05)

}
