% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{clusterAPAheatmap}
\alias{clusterAPAheatmap}
\title{Heatmap of cluster-specific APA genes}
\usage{
clusterAPAheatmap(ECoeffSig_Mat, FDR_P_cutoff = 0.05, CoeffCutoff = log(2))
}
\arguments{
\item{ECoeffSig_Mat}{significant APA table returned from \code{APAtest}.}

\item{FDR_P_cutoff}{The cutoff for FDR-controlled P values of Wald tests. Default to 0.05.}

\item{CoeffCutoff}{The cutoff for estimated coefficients of logistic regression.}
}
\value{
\code{clusterAPAheatmap} returns a heatmap of cluster-specific APA genes
}
\description{
\code{clusterAPAheatmap} draws heatmap of cluster-specific APA genes to show the degree and direction of APA. This function first selects genes that are significant cluster-specific APA in any cluster. Then, it draws heatmap with estimated coefficients from logistic regression.
}
\examples{
clusterAPAheatmap(ECoeffSig_Mat = result_from_estimateSig$ECoeffSig_Mat, FDR_P_cutoff=0.05, CoeffCutoff=log(2))

}
