% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSig.R
\name{IdentifyClusterAPA}
\alias{IdentifyClusterAPA}
\title{Identify gene-cluster-specific 3'UTR shortening or lengthening}
\usage{
IdentifyClusterAPA(ECoeffSig_Mat, WaldP_cutoff = 0.05, CoeffCutoff = log(2))
}
\arguments{
\item{ECoeffSig_Mat}{significant APA table returned from \code{APAtest}.}

\item{WaldP_cutoff}{The cutoff for P values of Wald tests. Default to 0.05.}

\item{CoeffCutoff}{The cutoff for estimated coefficients of logistic regression. Default to log(2).}
}
\value{
\code{IdentifyClusterAPA} returns a list consists of tables for every cluster. Tables contain the gene IDs, APA event IDs of 3'UTR shortening or lengthening that passed filters.
}
\description{
\code{IdentifyClusterAPA} Identifies gene-cluster-specific 3'UTR shortening or lengthening based on ECoeffSig_Mat output from \code{APAtest}.
}
\details{
After identifying APA genes, scMAPA can identify which cluster contribute to the significant APA dynamic across clusters and estimate what degree, what direction and the significance of the cluster-specific APA event. There are two criteria to be met for a gene to be considered as gene-cluster-specific APA gene. 1) Gene with P values of Wald test on a certain cluster smaller than WaldP_cutoff. 2) Gene with absoluste estimated coeffcient for same cluster greater than CoeffCutoff. Direction will be assigned based on sign of coefficients.
}
\examples{
result_list  <- IdentifyClusterAPA(ECoeffSig_Mat = sigAPA_table, WaldP_cutoff=0.05, CoeffCutoff=log(2))

}
