% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_filter.R
\name{readinPAsites}
\alias{readinPAsites}
\title{Generate filtered isofrom-specific count matrix}
\usage{
readinPAsites(
  path,
  NAcutoff = 3,
  CPMcutoff_L = 10,
  CPMcutoff_S = 10,
  clusterOfInterests = "all"
)
}
\arguments{
\item{path}{Path to the folder containing (only) outputs from DaPars for all clusters.}

\item{NAcutoff}{The lower limit for number of clusters in which the expression of a gene is detected (not NA).}

\item{CPMcutoff_L}{The gene-wise cutoff for average of CPM values of long isoforms across clusters. Default value is 10.}

\item{CPMcutoff_S}{The gene-wise cutoff for average of CPM values of short isoforms across clusters. Default value is 10.}

\item{clusterOfInterests}{The clusters should be included in the downstream analysis. It should be "all" to include all clusters or a vector of integers indicating the index of clusters of interest. At least 2 clusters should be provided.}
}
\value{
The return will be an isoform specific count matrix based on the DaPars2 estimation.
}
\description{
Read in the output from DaPars2, select clusters of interests, filter out genes with low quality, and generate a isoform-specific count matrix for APA detection.
}
\details{
\code{readinPAsites} will read in the output from BAMprocess programs and do selection and filtering based on parameters described above.

To filter out genes with low quality, only genes that expressed in number of clusters >= NAcutoff will be kept. \code{readinPAsites} will calculate CPM for long and short Isoforms separately by CPM=(raw count)*10^6/lib_size. Then, only genes with rowMeans(CPM of Long isoform)>CPMcutoff_L & rowMeans(CPM of Short isoform)>CPMcutoff_S will be kept. The flexibility of setting different cutoffs for long and short isoforms allows user to customize the QC by the feature of their samples.
}
\examples{
ISOMatrix  <- readinPAsites(path=getwd(), NAcutoff=3, CPMcutoff_L = 10, CPMcutoff_S = 10, clusterOfInterests = "all")
ISOMatrix  <- readin(path="Directory/storing/result/tables/from/previous/step")


}
