% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSig.R
\name{IdentifyClusterAPA}
\alias{IdentifyClusterAPA}
\title{Identify gene-cluster-specific 3'UTR shortening or lengthening}
\usage{
IdentifyClusterAPA(ECoeffSig_Mat, FDR_P_cutoff = 0.05, CoeffCutoff = log(2))
}
\arguments{
\item{ECoeffSig_Mat}{Matrix containing estimated coefficients, unadjusted P values of Wald tests on coefficients, and SE returned by model mode \code{estimateSig}.}

\item{FDR_P_cutoff}{The cutoff for FDR-controlled P values of Wald tests. Default to 0.05. Only cluster-specific APA events with FDR-controlled P values of Wald test smaller than this number will be considered as 3'UTR lengthening or shortening.}

\item{CoeffCutoff}{The cutoff for estimated coefficients of logistic regression. Default to log(2). Only cluster-specific APA events with absoluste value of estimated coeffcients greater than this cutoff will be considered as 3'UTR lengthening or shortening.}
}
\value{
\code{IdentifyClusterAPA} returns a list consists of a table for each cluster. Tables contain the 3'UTR shortening or lengthening passed filters and their gene IDs, APA event IDs.
}
\description{
\code{IdentifyClusterAPA} Identifies gene-cluster-specific 3'UTR shortening or lengthening based on ECoeffSig_Mat output from model mode \code{estimateSig}.
}
\examples{
result_list  <- IdentifyClusterAPA(ECoeffSig_Mat = result_from_estimateSig$ECoeffSig_Mat, FDR_P_cutoff=0.05, CoeffCutoff=log(2))

}
